ActivatablesManager and little cleanup
